Use host.docker.internal inside gnomock container
When possible (mac and windows), use host.docker.internal to access host
from inside gnomock container.
